Hack for gcc 4.5 and older
GCC 4.5 and older seem to have troubles accessing members of an
inner-class of a template class.  We need to call sizeof() of such a
member.  As a workaround, we declare a typedef for that member's type
and call sizeof() on the typedef instead of the inner-class member.

The same code used to compile fine with GCC 4.6 and older.

Signed-off-by: David Wagner <david.wagner@intel.com>