Merge amixer and alsactl capabilities and remove templating
BZ: 200396

This is a refactoring and simplification of all alsa plugins.

'mixer control' and 'device opening' capabilities used to be split in amixer
and alsactl plugin types, leading to a convoluted templating and inheritance
model.  It serves little purpose to split the two capabilities and merging them
allows a much simpler model.

Both the 'legacy alsa' and the 'tinyalsa' plugins derive from common base
classes.

Warning: Instead of "ALSACTL" and "AMIXER" plugin types, there is now a single
"ALSA" type (in Subsystem Structure files).  Existing Structure files need to
be updated.

Change-Id: I1f8fd29cbb453e8305d62ee705b437c0a3e7f1c0
Signed-off-by: David Wagner <david.wagner@intel.com>
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>