Use standard library path for pfw plugin location
The (tiny)alsa plugin used to be installed in
/system/lib/parameter-framework-plugins/Audio.
Since the Parameter-framework is now able to load plugins from
standard library locations such as /system/lib(64)/,
there is no need for this specific location anymore.

This patch moves the install directory of lib(tiny)alsa-subsystem.so
to the standard library path.

Issue: GMINL-1748

Change-Id: Ia560abd652faf8c190a9b4ac25677f3e929dd623
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>