LegacyAmixerControl: fix bytes control support mapped to component type
When bytes control is mapped to component type, PFW scalar size is the whole
size of amixer bytes control, this does not match scalar size of a byte.
Thus, the call of to/from Blackboard() calls can not be used.

To handle this case, use blackboardWrite() and blackboardRead() for hw access.

Signed-off-by: Miguel Gaio <miguel.gaio@intel.com>